Midway through conversion to integer ids and array storage for nodes.

+++ saginaw/src/main/java/project/search/ConcurrentBiDijkstra.java
@@ -51,7 +51,7 @@ public class ConcurrentBiDijkstra implements Searcher {
 
     }
 
-    public ArrayList<Long> search(long startNode, long endNode){
+    public ArrayList<Long> search(int startNode, int endNode){
 
         exploredA = 0;
         exploredB = 0;
