+++ saginaw/src/main/java/project/search/Searcher.java
@@ -3,7 +3,7 @@ package project.search;
 import java.util.ArrayList;
 
 public interface Searcher {
-    ArrayList<Long> search(long src, long dst);
+    ArrayList<Long> search(int src, int dst);
     double getDist();
     int getExplored();
 }
